--- a/.github/workflows/benchmark.yml
+++ b/.github/workflows/benchmark.yml
@@ -1,0 +1,46 @@
+name: Benchmark
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+  workflow_dispatch:
+
+jobs:
+  benchmark:
+    if: ${{ !contains(github.event.head_commit.message, '[ci skip]') }}
+    runs-on: ubuntu-latest
+    env:
+      BUNDLE_JOBS: 4
+      BUNDLE_RETRY: 3
+      CI: true
+      RAILS_VERSION: ${{ matrix.rails_version }}
+      DATABASE_URL: postgresql://postgres:postgres@localhost:5432
+    services:
+      redis:
+        image: redis:7.0-alpine
+        ports: ["6379:6379"]
+        options: --health-cmd="redis-cli ping" --health-interval 1s --health-timeout 3s --health-retries 30
+    strategy:
+      fail-fast: false
+      matrix:
+        rails_version: ["~> 7.0", "~> 8.0.0"]
+        adapter: ["redis", "async"]
+        features: [""]
+        include:
+        - rails_version: "~> 8.0.0"
+          adapter: "redis"
+          features: "FASTLANE_BROADCASTS=1"
+    steps:
+      - uses: actions/checkout@v4
+      - uses: ruby/setup-ruby@v1
+        with:
+          ruby-version: 3.3
+          bundler-cache: true
+      - name: Run benchmarks
+        env:
+          ACTION_CABLE_ADAPTER: ${{ matrix.adapter }}
+          MODE: smoke
+        run: |
+          ${{ matrix.features }} bundle exec ruby --yjit benchmarks/broadcasting.rb