--- a/lib/action_cable/connection/internal_channel.rb
+++ b/lib/action_cable/connection/internal_channel.rb
@@ -18,7 +18,7 @@ module ActionCable
 
         def subscribe_to_internal_channel
           if connection_identifier.present?
-            callback = -> (message) { process_internal_message ActiveSupport::JSON.decode(message) }
+            callback = -> (message) { process_internal_message ActiveSupport::JSON.decode(message.data) }
             @_internal_subscriptions ||= []
             @_internal_subscriptions << [ internal_channel, callback ]